requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.2
decorator==5.1.1
Django==5.0.4
django-celery-beat==2.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipython==8.23.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
msgpack==1.0.8
multidict==6.0.5
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.35
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
twilio==9.0.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zope.interface==6.3
